requirements.txt
pygments~=2.11.2
pip~=21.2.4
toml~=0.10.2
wheel~=0.37.1
cryptography~=36.0.2
lxml~=4.8.0
tornado~=6.1
certifi~=2021.10.8
setuptools~=58.0.4
streamlit~=1.8.1
requests~=2.27.1
pandas~=1.3.5
instagramy~=4.5
bs4~=0.0.1
beautifulsoup4~=4.10.0
pillow~=9.1.0
selenium~=4.1.3